Output枠のSendMessageの結果はNET系は、結果をLRESULTではなくIntに収める形とした。32/64共通の数値の範囲のはずなので、まずまちがいなくintを超えないはず。

+++ win-hidemaru-app/hm_dotnet/hm.NET.src/hmNETStaticLib/参照用のインターフェイス/HmReferenceInterface.cs
@@ -281,7 +281,7 @@ namespace Hidemaru
                 }
             }
 
-            public static IntPtr SendMessage(int command_id)
+            public static int SendMessage(int command_id)
             {
                 return hmNETDynamicLib.Hidemaru.OutputPane.SendMessge(command_id);
             }
